# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GSDB API",
    "version": "0.1.0",
    "description": "\n# GSDB API Documentation\n\n**Authentication:** To use the API, you must provide an API key as a Bearer token.  \nThis token can be obtained in your user profile edit page on the website: [User Area](https://gsdb-web.onrender.com/user-area).  \nInclude it in the header of every request:\n\n`Authorization: Bearer <API_KEY>`\n\nFor any other information, like how to use query parameters or the different database fields, you can visit the [GSDB Wiki](https://github.com/Game-Save-DataBase/GSDB_Server/wiki)\n"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_KEY"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/savedatas": {
      "get": {
        "summary": "Search savedatas",
        "description": "Search savedatas by query filters. Supports MongoDB query operands. Can fetch by ID or other fields.",
        "tags": [
          "Savedatas"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "description": "Savedata ID (optional)"
          },
          {
            "in": "query",
            "name": "title",
            "schema": {
              "type": "string"
            },
            "description": "Title of savedata (optional)"
          },
          {
            "in": "query",
            "name": "gameID",
            "schema": {
              "type": "number"
            },
            "description": "Game ID (optional)"
          },
          {
            "in": "query",
            "name": "userID",
            "schema": {
              "type": "string"
            },
            "description": "User ID (optional)"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of savedata objects or single savedata"
          },
          "204": {
            "description": "No coincidences found"
          },
          "400": {
            "description": "Invalid query"
          }
        }
      },
      "post": {
        "summary": "Upload a savedata",
        "description": "Upload a savedata file with optional screenshots. Requires authentication.",
        "tags": [
          "Savedatas"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "screenshots": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "gameID": {
                    "type": "number"
                  },
                  "platformID": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "tagID": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Savedata uploaded successfully"
          },
          "400": {
            "description": "Invalid upload or missing fields"
          }
        }
      },
      "put": {
        "summary": "Update savedata",
        "description": "Update savedata fields. Authentication required.",
        "tags": [
          "Savedatas"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Savedata ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Fields to update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Savedata updated successfully"
          },
          "400": {
            "description": "Invalid fields or missing ID"
          },
          "404": {
            "description": "Savedata not found"
          }
        }
      },
      "delete": {
        "summary": "Delete savedata",
        "description": "Deletes a savedata by ID. Authentication required.",
        "tags": [
          "Savedatas"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Savedata deleted successfully"
          },
          "404": {
            "description": "Savedata not found"
          }
        }
      }
    },
    "/savedatas/search": {
      "get": {
        "summary": "Search savedatas by text",
        "description": "Full-text search across title, description, game title, platform, and user info.",
        "tags": [
          "Savedatas"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "Search term"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Limit number of results"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer"
            },
            "description": "Offset for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of matching savedata"
          },
          "204": {
            "description": "No coincidences found"
          }
        }
      }
    },
    "/savedatas/async": {
      "post": {
        "summary": "Async upload a savedata",
        "description": "Starts asynchronous upload and processing of a savedata file. Returns immediately.",
        "tags": [
          "Savedatas"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "screenshots": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "gameID": {
                    "type": "number"
                  },
                  "platformID": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "tagID": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Asynchronous process started"
          },
          "400": {
            "description": "Invalid upload or missing fields"
          }
        }
      }
    },
    "/savedatas/update-rating": {
      "put": {
        "summary": "Like or dislike a savedata",
        "description": "Updates likes or dislikes for a savedata by the authenticated user.",
        "tags": [
          "Savedatas"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "mode",
            "schema": {
              "type": "string",
              "enum": [
                "like",
                "dislike"
              ]
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Rating updated successfully"
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "404": {
            "description": "Savedata not found"
          }
        }
      }
    },
    "/savedatas/reset-rating": {
      "put": {
        "summary": "Reset user rating on a savedata",
        "description": "Removes the like or dislike of the authenticated user for a specific savedata.",
        "tags": [
          "Savedatas"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ratings reset successfully"
          },
          "404": {
            "description": "Savedata not found"
          }
        }
      }
    },
    "/savedatas/dev/wipe": {
      "delete": {
        "summary": "Wipe all savedatas",
        "description": "Deletes all savedatas and resets related game upload counters. Development mode only.",
        "tags": [
          "Savedatas"
        ],
        "responses": {
          "200": {
            "description": "All savedatas deleted successfully"
          },
          "500": {
            "description": "Server error while wiping savedatas"
          }
        }
      }
    },
...
    "/assets/savedata/{id}": {
      "get": {
        "summary": "Download the main savefile",
        "description": "Returns the primary zip file for a savefile. Increments download count.",
        "tags": [
          "Assets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Savefile ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Savefile zip"
          },
          "404": {
            "description": "Savefile not found"
          }
        }
      }
    },
...
  },
...
}